Bump version; git-ize Release in specfile
Bump version to 0.8.1 so we can cleanly change the Release field in
the specfile (and resulting RPMs, of course) to include git-related
history information (similar to Zathras).  Git metadata is now used to
determine the latest commit hash, the "modified" state of the tree,
and the number of commits since the most recent tag.